+++ docs/source/conf.py
@@ -13,13 +13,14 @@
 
 import os
 import sys
-sys.path.insert(0, os.path.abspath('../../src/'))
+
+sys.path.insert(0, os.path.abspath("../../src/"))
 
 # -- Project information -----------------------------------------------------
 
-project = 'festo-cpx-io'
-copyright = '2023, Festo SE & Co. KG'
-author = 'Martin Wiesner, Elias Rosch'
+project = "festo-cpx-io"
+copyright = "2023, Festo SE & Co. KG"
+author = "Martin Wiesner, Elias Rosch"
 
 
 # -- General configuration ---------------------------------------------------
@@ -33,10 +34,10 @@ extensions = [
     "sphinx_rtd_theme",  # Use Read-the-Docs theme
     "myst_parser",  # Use markdown files in the documentation along restructured text
 ]
-source_suffix = ['.rst', '.md']
+source_suffix = [".rst", ".md"]
 
 # Add any paths that contain templates here, relative to this directory.
-templates_path = ['_templates']
+templates_path = ["_templates"]
 
 # List of patterns, relative to source directory, that match files and
 # directories to ignore when looking for source files.
@@ -49,15 +50,19 @@ exclude_patterns = []
 # The theme to use for HTML and HTML Help pages.  See the documentation for
 # a list of builtin themes.
 #
-html_theme = 'sphinx_rtd_theme'
+html_theme = "sphinx_rtd_theme"
 
 # Add any paths that contain custom static files (such as style sheets) here,
 # relative to this directory. They are copied after the builtin static files,
 # so a file named "default.css" will overwrite the builtin "default.css".
 # html_static_path = ['_static']
 
-autoclass_content = 'both'
+autoclass_content = "both"
 autodoc_default_options = {
     # 'inherited-members': True,
-    'member-order': 'bysource',
+    "members": True,  # Include members (functions, methods)
+    "member-order": "bysource",
+    # "special-members": True,  # Include special members (e.g., __init__)
+    "undoc-members": True,  # Include members without docstrings
+    # "private-members": True,  # Include private members (with leading underscore)
 }