+++ src/cpx_io/cpx_system/cpx_e/e8do.py
@@ -6,6 +6,7 @@
 from cpx_io.cpx_system.cpx_base import CpxBase
 from cpx_io.cpx_system.cpx_e.cpx_e_module import CpxEModule
 from cpx_io.utils.boollist import int_to_boollist, boollist_to_int
+from cpx_io.utils.logging import Logging
 
 
 class CpxE8Do(CpxEModule):
@@ -25,32 +26,50 @@ class CpxE8Do(CpxEModule):
 
     @CpxBase.require_base
     def read_channels(self) -> list[bool]:
-        """read all channels as a list of bool values"""
+        """read all channels as a list of bool values
+
+        :return: Values of all channels
+        :rtype: list[bool]
+        """
         data = self.base.read_reg_data(self.input_register)[0]
-        return int_to_boollist(data, num_bytes=1)
+        ret = int_to_boollist(data, num_bytes=1)
+        Logging.logger.info(f"{self.name}: Reading channels: {ret}")
+        return ret
+
+    @CpxBase.require_base
+    def read_channel(self, channel: int) -> bool:
+        """read back the value of one channel
+
+        :param channel: Channel number, starting with 0
+        :type channel: int
+        :return: Value of the channel
+        :rtype: bool
+        """
+        return self.read_channels()[channel]
 
     @CpxBase.require_base
     def write_channels(self, data: list[bool]) -> None:
-        """write all channels with a list of bool values"""
+        """write all channels with a list of bool values
+
+        :param data: list of bool values containing exactly 4 elements for each output channel
+        :type data: list[bool]
+        """
         if len(data) != 8:
             raise ValueError(f"Data len error: expected: 8, got: {len(data)}")
         integer_data = boollist_to_int(data)
         self.base.write_reg_data(integer_data, self.output_register)
 
-    @CpxBase.require_base
-    def read_status(self) -> list[bool]:
-        """read module status register. Further information see module datasheet"""
-        data = self.base.read_reg_data(self.input_register + 1)[0]
-        return int_to_boollist(data, num_bytes=2)
-
-    @CpxBase.require_base
-    def read_channel(self, channel: int) -> bool:
-        """read back the value of one channel"""
-        return self.read_channels()[channel]
+        Logging.logger.info(f"{self.name}: Setting channels to {data}")
 
     @CpxBase.require_base
     def write_channel(self, channel: int, value: bool) -> None:
-        """set one channel to logic value"""
+        """set one channel to logic value
+
+        :param channel: Channel number, starting with 0
+        :type channel: int
+        :value: Value that should be written to the channel
+        :type value: bool
+        """
         data = self.base.read_reg_data(self.input_register)[0]  # read current value
         mask = 1 << channel  # Compute mask, an integer with just bit 'channel' set.
         data &= ~mask  # Clear the bit indicated by the mask
@@ -59,19 +78,41 @@ class CpxE8Do(CpxEModule):
 
         self.base.write_reg_data(data, self.output_register)
 
+        Logging.logger.info(f"{self.name}: Setting channel {channel} to {value}")
+
+    @CpxBase.require_base
+    def read_status(self) -> list[bool]:
+        """read module status register. Further information see module datasheet
+
+        :return: status information (see datasheet)
+        :rtype: list[bool]"""
+        data = self.base.read_reg_data(self.input_register + 1)[0]
+        ret = int_to_boollist(data, 2)
+        Logging.logger.info(f"{self.name}: Reading status: {ret}")
+        return ret
+
     @CpxBase.require_base
     def set_channel(self, channel: int) -> None:
-        """set one channel to logic high level"""
+        """set one channel to logic high level
+
+        :param channel: Channel number, starting with 0
+        :type channel: int"""
         self.write_channel(channel, True)
 
     @CpxBase.require_base
     def clear_channel(self, channel: int) -> None:
-        """set one channel to logic low level"""
+        """set one channel to logic low level
+
+        :param channel: Channel number, starting with 0
+        :type channel: int"""
         self.write_channel(channel, False)
 
     @CpxBase.require_base
     def toggle_channel(self, channel: int) -> None:
-        """set one channel the inverted of current logic level"""
+        """set one channel the inverted of current logic level
+
+        :param channel: Channel number, starting with 0
+        :type channel: int"""
         data = (
             self.base.read_reg_data(self.input_register)[0] & 1 << channel
         ) >> channel
@@ -83,7 +124,9 @@ class CpxE8Do(CpxEModule):
             raise ValueError(f"Value {data} must be between 0 and 1")
 
     @CpxBase.require_base
-    def configure_diagnostics(self, short_circuit=None, undervoltage=None):
+    def configure_diagnostics(
+        self, short_circuit: bool = None, undervoltage: bool = None
+    ) -> None:
         """
         The "Diagnostics of short circuit at output" parameter defines whether
         the diagnostics of the outputs in regard to short circuit or
@@ -93,6 +136,11 @@ class CpxE8Do(CpxEModule):
         deactivated with regard to undervoltage.
         When the diagnostics are activated,
         the error will be sent to the bus module and displayed on the module by the error LED.
+
+        :param short_circuit: diagnostics of short circuit
+        :type short_circuit: bool
+        :param undervoltage: diagnostics of undervoltage
+        :type undervoltage: bool
         """
         function_number = 4828 + 64 * self.position + 0
         reg = self.base.read_function_number(function_number)
@@ -109,6 +157,10 @@ class CpxE8Do(CpxEModule):
         )
 
         self.base.write_function_number(function_number, value_to_write)
+        Logging.logger.info(
+            f"{self.name}: Writing diagnostics parameter short circuit: {short_circuit}, "
+            f"undervoltage: {undervoltage}"
+        )
 
     @CpxBase.require_base
     def configure_power_reset(self, value: bool):
@@ -119,6 +171,9 @@ class CpxE8Do(CpxEModule):
         In the case of the "Leave power switched off" setting,
         the CPX-E automation system must be switched off and on or
         the corresponding output must be reset and to restore the power.
+
+        :param value: behaviour after power reset
+        :type value: bool
         """
         function_number = 4828 + 64 * self.position + 1
         reg = self.base.read_function_number(function_number)
@@ -130,3 +185,7 @@ class CpxE8Do(CpxEModule):
             value_to_write = reg & 0xFD
 
         self.base.write_function_number(function_number, value_to_write)
+
+        Logging.logger.info(
+            f"{self.name}: Setting behaviour after power reset to {value}"
+        )